Move admin arrays to shared module
